+++ system/container/container.go
@@ -40,6 +40,9 @@ var engineChecked bool
 // engineCache cached engine info
 var engineCache string
 
+// isK8s is set to true if we inside K8s pod container
+var isK8s bool
+
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // GetEngine returns container engine name if used
@@ -65,11 +68,18 @@ func IsContainer() bool {
 	return GetEngine() != ""
 }
 
+// IsK8s returns true if we are inside K8s pod container
+func IsK8s() bool {
+	return GetEngine() != "" && isK8s
+}
+
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // guessEngine tries to guess container engine based on information from /proc/1/mounts
 func guessEngine(mountsData string) string {
 	_, err := os.Stat(dockerEnv)
+
+	isK8s = strings.Contains(mountsData, "kubernetes.io")
 
 	switch {
 	case strings.Contains(mountsData, "overlay-container /function/code/rootfs"):